Refactor: Add `_stbt.imgutils` containing common image utility functions
This is part of an effort to split up `core.py` into smaller chunks.

This is the minimum set of functions that are used by core.py and the
`xxx_motion` functions.  I think it makes sense to put a function in the
most specific file possible.  This is in preparation for moving the motion
functions into `motion.py`.
